@@ -1,8 +1,8 @@
 package com.xzwzz.fastcamera.delegate;
 
 
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentManager;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
 
 /**
  * @author xzwzz
